test(ops): move A5 MC2 dispatch quant guard to e2e single-card A5
quant_mode=4/y_dtype is internal to get_dispatch_mc2_kwargs and not
observable from model outputs, so it cannot be a model-level e2e check. Move
the A5 MXFP dispatch assertion to an e2e single-card A5 test that runs the
dispatcher method on real A5 (real device detection, A5-gated, skips on
non-A5) and drop the CPU-mocked UT. Register the e2e file + estimated_times.
